refactor: 공통 타입을 global 타입으로 변경
