--- src/@types/index.d.ts
@@ -1,8 +1,0 @@
-declare namespace Common {
-  interface Response<T> {
-    result?: string;
-    message?: string;
-    data: T;
-  }
-  type Modify<T, R> = Omit<T, keyof R> & R;
-}
