+++ src/apis/review/apis.ts
@@ -3,7 +3,7 @@ import ky from 'ky';
 import type { Review } from './types';
 
 export const getReviews = async (placeId: number, size: number, pageParam: number) => {
-  const { data } = await ky.get('/db/reviews.json').json<Common.Response<{ reviews: Review[] }>>();
+  const { data } = await ky.get('/db/reviews.json').json<APIResponse<{ reviews: Review[] }>>();
 
   return {
     data,
@@ -12,7 +12,7 @@ export const getReviews = async (placeId: number, size: number, pageParam: numbe
 };
 
 export const getMemberReviews = async (memberId: number, size: number, pageParam: number) => {
-  const { data } = await ky.get('/db/reviews.json').json<Common.Response<{ reviews: Review[] }>>();
+  const { data } = await ky.get('/db/reviews.json').json<APIResponse<{ reviews: Review[] }>>();
 
   return {
     data,
