+++ src/hooks/useGeolocation.ts
@@ -1,7 +1,5 @@
 import { useState, useEffect } from 'react';
 
-import type { Coordinates } from '@/types/coordinates';
-
 type locationType = {
   loaded: boolean;
   coordinates?: Coordinates;
